Move creation of new change request form submissions into its own component file
